Add template file for TodoList item
